// play/chess-tourney-frontend/src/api/model/playerColor.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */

export type PlayerColor = (typeof PlayerColor)[keyof typeof PlayerColor];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PlayerColor = {
  White: "White",
  Grey: "Grey",
  Black: "Black",
} as const;